Update vite.md
There was some missing steps that update the main.ts file to add references to the Tailwind css file. This would then break the Shadcn UI when the project was run with npm run dev. I went through the docs half a dozen times and could never get the project to work correctly until referencing some stackblitz examples that had the correct configuration for Vite